Move app to containers
