+++ src/index.js
@@ -4,7 +4,7 @@ import registerServiceWorker from './registerServiceWorker';
 import { Provider } from 'react-redux';
 import store from './redux/store';
 
-import App from './components/App';
+import App from './containers/App';
 
 const app = (
   <Provider store={store}>
